Refactor: Extract new mod `async_executor`
